Associated Tips with Troubleshooting Exceptions: System.BadImageFormatException
If your application uses 32-bit components, make sure that it always runs as a 32-bit application.
If the Platform target property for your application project is set to AnyCPU, the compiled application can be run in either 64-bit or 32-bit mode. When it runs as a 64-bit application, the just-in-time (JIT) compiler produces 64-bit native code. If the application depends on a 32-bit managed or unmanaged component, that component will fail to load in 64-bit mode. To correct this problem, set the project's Platform target property to x86 and recompile.

Using 64-bit components in our project
Current Directory is modified